extracting note selection handling from artist
