+++ src/views/music/Stave.js
@@ -1,8 +1,10 @@
 // @flow
 
 import React, { Component } from 'react'
-import { Flow } from 'vexflow'
+import { connect } from 'react-redux'
 import _ from 'lodash'
+import { Flow } from 'vexflow'
+import { studentAddedNote } from 'modules/student-test/actions'
 import type { Question as QuestionType } from 'modules/student-test'
 
 import VexTab from 'modules/music/notation/vextab'
@@ -22,7 +24,7 @@ type Props = {
   disableEditing?: boolean,
 }
 
-export default class Stave extends Component {
+class Stave extends Component {
   staveContainer: React.Element<any>
   artist: Artist
   props: Props
@@ -40,7 +42,6 @@ export default class Stave extends Component {
       annotations = '',
       notes = '',
       question = {},
-      disableEditing = false,
     } = this.props
 
     const text = (annotations ? `text ${annotations}` : '')
@@ -52,12 +53,38 @@ export default class Stave extends Component {
       notes ${notes.trim()}
       ${text}
     `
-    this.artist = new Artist(10, 10, width, { scale: 1, disableEditing }, question)
+    this.artist = new Artist(10, 10, width, {
+      scale: 1,
+      addNote: (position, pitch) => this.addStudentNote(position, pitch)
+    }, question)
+    
     const vextab = new VexTab(this.artist)
     const renderer = new Flow.Renderer(this.staveContainer, Flow.Renderer.Backends.SVG)
     vextab.parse(notation)
 
     this.artist.render(renderer)
+  }
+
+  addStudentNote(position, pitch) {
+    const studentNotes = this.props.question.student
+    const currentPosition = position
+    const studentNotesAtPosition = _
+      .filter(studentNotes, x => x.position === currentPosition)
+      .length
+
+    const maxNumberOfNotes = (
+      this.artist.question.options &&
+      this.artist.question.options.maxNotesPerMeasure) ||
+      -1
+
+    if (maxNumberOfNotes === -1 || studentNotesAtPosition < maxNumberOfNotes) {
+      this.props.studentAddedNote({
+        pitch: pitch,
+        duration: 'q',
+        position: currentPosition,
+        questionId: this.props.question.id,
+      })
+    }
   }
 
   render(): React.Element<any> {
@@ -72,3 +99,16 @@ export default class Stave extends Component {
     )
   }
 }
+
+
+const mapStateToProps = (state, ownProps) => {
+  return { }
+}
+
+const mapDispatchToProps = (dispatch: Dispatch<any>) => {
+  return {
+    studentAddedNote: ((note) => dispatch(studentAddedNote(note))),
+  }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Stave)
